pyproject.toml
[project]
name = "local-rag-chat"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "Jarrett Meyer", email = "jarrettmeyer@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "psycopg-pool>=3.2.6",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "typer>=0.16.1",
]

[project.scripts]
local-rag-chat = "local_rag_chat.cli:app"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
